Revert "Dh apmspii 874 draft implementation sync wrap"
