+++ azure/templates/pds-tests.yml
@@ -11,17 +11,22 @@ steps:
 
   - bash: |
       set -euo pipefail
+
       if [ -d $(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)/tests ]; then
         echo "Running ${{ parameters.test_type }} tests"
+
         # Go to tests folder
         cd $(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)/tests
+
         # Install dependencies
         npm install
+
         # Run tests
         echo "node_modules/.bin/node --unhandled-rejections=strict e2e/test-runner.js $(IDP_URL) $(API_URL) ${{ parameters.test_pack_path }}  ./e2e/environments/deploy.postman-environment.json"
         node_modules/.bin/node --unhandled-rejections=strict e2e/test-runner.js $(IDP_URL) $(API_URL) ${{ parameters.test_pack_path }}  ./e2e/environments/deploy.postman-environment.json;
       fi
     displayName: 'Run ${{ parameters.test_type }} tests'
+
   - task: PublishTestResults@2
     displayName: 'Publish ${{ parameters.test_type }} Test Results'
     inputs:
@@ -42,10 +47,11 @@ steps:
       export PDS_BASE_PATH="$(SERVICE_BASE_PATH)"
       export APIGEE_ENVIRONMENT="$(ENVIRONMENT)"
       export KEY_ID="test-1"
-
+      
       poetry run pytest -v tests/functional/test_application_restricted.py --junitxml=tests/functional-test-report.xml
     displayName: Run application restricted tests
     workingDirectory: "$(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)"
+
   - task: PublishTestResults@2
     displayName: 'Publish application restricted test results'
     inputs:
@@ -63,7 +69,7 @@ steps:
       export APIGEE_API_TOKEN="$(secret.AccessToken)"
       export OAUTH_BASE_URI="https://$(ENVIRONMENT).api.service.nhs.uk"
       export OAUTH_PROXY="oauth2"
-
+ 
       poetry run pytest -v tests/functional/test_proxy_behaviour.py --junitxml=tests/proxy-behaviour-test-report.xml
     displayName: Run proxy behaviour tests
     workingDirectory: "$(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)"
@@ -85,32 +91,9 @@ steps:
       poetry run pytest -v tests/user_restricted/test_user_restricted.py --junitxml=tests/user-restricted-test-report.xml
     displayName: Run user restricted tests
     workingDirectory: "$(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)"
+
   - task: PublishTestResults@2
     displayName: 'Publish user restricted test results'
     inputs:
       testResultsFiles: $(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)/tests/user-restricted-test-report.xml
-      failTaskOnFailedTests: true
-
-  - bash: |
-      export APPLICATION_RESTRICTED_SIGNING_KEY_PATH="$(Pipeline.Workspace)/secrets/$(JWT_TESTING_PRIVATE_KEY)"
-      export APPLICATION_RESTRICTED_WITH_ASID_SIGNING_KEY_PATH="$(Pipeline.Workspace)/secrets/$(JWT_TESTING_WITH_ASID_PRIVATE_KEY)"
-      export APPLICATION_RESTRICTED_API_KEY=" "
-      export APPLICATION_RESTRICTED_WITH_ASID_API_KEY=" "
-      export PDS_BASE_PATH="$(SERVICE_BASE_PATH)"
-      export APIGEE_ENVIRONMENT="$(ENVIRONMENT)"
-      export KEY_ID="test-1"
-      export APIGEE_API_TOKEN="$(secret.AccessToken)"
-      export OAUTH_BASE_URI="https://$(ENVIRONMENT).api.service.nhs.uk"
-      export OAUTH_PROXY="oauth2"
-      export CLIENT_ID="$(CLIENT_ID)"
-      export CLIENT_SECRET="$(CLIENT_SECRET)"
-
-      poetry run pytest -v tests/functional/test_sync_wrap.py --junitxml=tests/sync-wrap-test-report.xml
-    displayName: Run sync wrap tests
-    workingDirectory: "$(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)"
-
-  - task: PublishTestResults@2
-    displayName: "Publish sync wrap test results"
-    inputs:
-      testResultsFiles: $(Pipeline.Workspace)/s/$(SERVICE_NAME)/$(SERVICE_ARTIFACT_NAME)/tests/sync-wrap-test-report.xml
       failTaskOnFailedTests: true
