--- proxies/live/apiproxy/policies/AssignMessage.AddSyncWaitHeader.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<AssignMessage async="false" continueOnError="false" enabled="true" name="AssignMessage.AddSyncWaitHeader">
-  <Set>
-      <Headers>
-          <Header name="X-Sync-Wait">10</Header>
-      </Headers>
-  </Set>
-  <IgnoreUnresolvedVariables>false</IgnoreUnresolvedVariables>
-  <AssignTo createNew="false" transport="http" type="request"/>
-</AssignMessage>
