--- proxies/live/apiproxy/policies/AssignMessage.Errors.ServiceUnavailable.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<AssignMessage name="AssignMessage.Errors.ServiceUnavailable">
-  <Set>
-    <StatusCode>503</StatusCode>
-    <ReasonPhrase>Service unavailable</ReasonPhrase>
-    <Headers>
-      <Header name="Retry-After">5</Header>
-    </Headers>
-  </Set>
-  <AssignVariable>
-    <Name>pds.error.code</Name>
-    <Value>timeout</Value>
-  </AssignVariable>
-  <AssignVariable>
-    <Name>pds.error.coding.code</Name>
-    <Value>SERVICE_UNAVAILABLE</Value>
-  </AssignVariable>
-  <AssignVariable>
-    <Name>pds.error.coding.display</Name>
-    <Value>Service unavailable</Value>
-  </AssignVariable>
-  <AssignVariable>
-    <Name>pds.error.diagnostics</Name>
-    <Value>
-      The downstream domain processing has not completed within the configured timeout period.
-      Retry your request after the time specified by the 'Retry-After' header, using the same 'X-Request-ID'.
-    </Value>
-  </AssignVariable>
-  <IgnoreUnresolvedVariables>true</IgnoreUnresolvedVariables>
-</AssignMessage>
