--- proxies/live/apiproxy/policies/AssignMessage.SetContentLocationAbsoluteUrl.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<AssignMessage async="false" continueOnError="false" enabled="true" name="AssignMessage.SetContentLocationAbsoluteUrl">
-    <DisplayName>AssignMessage.SetContentLocationAbsoluteUrl</DisplayName>
-    <Set>
-        <Headers>
-            <Header name="Content-Location">https://{request.header.Host}/{proxy.basepath}{response.header.content-location}</Header>
-        </Headers>
-    </Set>
-    <IgnoreUnresolvedVariables>false</IgnoreUnresolvedVariables>
-    <AssignTo createNew="false" transport="http" type="response"/>
-</AssignMessage>
