+++ tests/e2e/user-restricted-smoke.json
@@ -255,7 +255,7 @@
 								"id": "94ce6628-5334-4717-b49e-fd8a36296155",
 								"exec": [
 									"//check the response status\r",
-									"pm.test.skip(\"Response has a status of 202\", function () {\r",
+									"pm.test(\"Response has a status of 202\", function () {\r",
 									"    pm.response.to.have.status(202);\r",
 									"})\r",
 									"\r",
@@ -263,10 +263,10 @@
 									"pm.environment.set('Content-Location', pm.response.headers.get(\"Content-Location\").substr(1, 46))\r",
 									"\r",
 									"//check that the correct Response Headers have been returned in the response\r",
-									"pm.test.skip(\"Response contains the correct X-Request-ID header\", function() {\r",
+									"pm.test(\"Response contains the correct X-Request-ID header\", function() {\r",
 									"    pm.expect(pm.request.headers.get(\"X-Request-ID\")).to.eql(pm.response.headers.get(\"X-Request-ID\"))\r",
 									"})\r",
-									"pm.test.skip(\"Response contains the correct X-Correlation-ID header\", function() {\r",
+									"pm.test(\"Response contains the correct X-Correlation-ID header\", function() {\r",
 									"    pm.expect(pm.request.headers.get(\"X-Correlation-ID\")).to.eql(pm.response.headers.get(\"X-Correlation-ID\"))\r",
 									"})"
 								],
@@ -336,12 +336,12 @@
 									"var response = pm.response.json();\r",
 									"\r",
 									"//check the response status\r",
-									"pm.test.skip(\"Response has a status of 200\", function () {\r",
+									"pm.test(\"Response has a status of 200\", function () {\r",
 									"    pm.response.to.have.status(200);\r",
 									"})\r",
 									"\r",
 									"//check the update has been made\r",
-									"pm.test.skip(\"Response contains updated patient record\", function() {\r",
+									"pm.test(\"Response contains updated patient record\", function() {\r",
 									"    pm.expect(response.birthDate).to.eql(pm.environment.get(\"birthDate\"))\r",
 									"})"
 								],
