+++ tests/functional/config_files/config.py
@@ -12,7 +12,3 @@ KEY_ID = ENV["key_id"]
 
 # PDS
 PDS_BASE_PATH = ENV["pds_base_path"]
-
-# App details
-CLIENT_ID = ENV['client_id']
-CLIENT_SECRET = ENV['client_secret']
