+++ tests/pytest.ini
@@ -4,10 +4,7 @@ markers =
 
     # JIRA tickets
     apmspii_627
-    apmspii_832
-    apmspii_874
 
     # 
     quota
     rate_limit
-    sync_wrap
