+++ tests/scripts/pds_request.py
@@ -16,12 +16,11 @@ class PdsRecord:
             self.headers = dict(response.headers.items())
             self.redirects = self._get_redirects(response)
             self.url = response.url
+
             try:
                 self.response = loads(response.text)
-            except JSONDecodeError as e:
-                print(f'UNEXPECTED RESPONSE: {response.text}. \n Error: {e}')
-                self.response = {}
-                # raise Exception(f'UNEXPECTED RESPONSE: {response.text}. \n Error: {e}')
+            except JSONDecodeError:
+                raise Exception(f'UNEXPECTED RESPONSE {response.text}:')
 
         # if the response is a list of entries i.e. a response from a search
         if 'entry' in self.response:
@@ -91,12 +90,12 @@ class PdsRecord:
 
 
 class GenericPdsRequestor(GenericRequest):
-    """A utility class to make generic PDS requests. """
+    """This class is used to make PDS requests. """
 
     def __init__(
         self,
-        pds_base_path: str,
-        base_url: str,
+        pds_base_path: str,  # contains fhir ext
+        base_url: str,  # From config per test
         token: str = None,
         headers: dict = None
     ):
@@ -112,40 +111,7 @@ class GenericPdsRequestor(GenericRequest):
                 'X-Request-ID': str(uuid4()),
             }
 
-    @property
-    def headers(self):
-        return self._headers
-
-    @headers.setter
-    def headers(self, headers: dict):
-        self._headers.update(headers)
-
     def get_patient_response(self, patient_id: str, **kwargs) -> PdsRecord:
         """Return a PDS record as an object"""
         response = self.get(f'{self.base_url}/Patient/{patient_id}', headers=self._headers, ** kwargs)
-        return PdsRecord(response)
-
-    def update_patient_response(self, patient_id: str, payload: dict, **kwargs) -> PdsRecord:
-        """Sends a patch request and returns a PDS record as an object.
-
-        Args:
-            patient_id (str): Patient to update.
-            payload: A JSON patch. E.g:
-                {
-                    "patches": [
-                        { "op": "add", "path": "/deceasedDate", "value": "2020-01-01" }
-                    ]
-                }
-            See for more details:
-        https://digital.nhs.uk/developer/api-catalogue/personal-demographics-service-fhir#api-Default-updatePatientPartial.
-
-        Returns:
-            response (PdsRecord): PDS Record object."""
-        response = self.patch(
-            f'{self.base_url}/Patient/{patient_id}',
-            headers=self._headers,
-            json=payload,
-            **kwargs
-        )
-
         return PdsRecord(response)
